Plumbing dynamic shape through SPIR-V CodeGen
This commit adjusts the SPIR-V pipeline to support dynamic
shape. An initial test that adds two tensor<?x4xf32> is
passing on Vulkan backend after this change. Specifically,
the changes made to the SPIR-V pipeline include:

1. This commit inlines the impl dispatch function into the
wrapper dispatch function at the very first step. This
eliminates the function split so that we can avoid the
complicity to convert across function boundaries.

2. This commit introduces a new iree.placeholder op that
one can use to get the interface buffer from a specific
descriptor set and binding and feed this placeholder op
to computation. It serves a similar purpose like function
arguments but has the form of an op.

3. The conversion from linalg on tensors to linalg on buffers
is simplified that we don't need the double conversion now.
Previously we need to first convert the impl function from
tensor to memref, and then in the impl function to load
the memref to get tensors before actually performing any
conversion to linalg. Now the tensor to memref conversion
only happens during linalg on tensors to linalg on buffers.

4. We don't create gpu.func and gpu.module any more. Instead
all the conversion happens on ops inside the same inlined
dispatch function. This simplifies the interface handling
because now we don't need to pass SPIR-V interface info
through gpu.func layer anymore. The result is that we only
use the processor ID ops from the gpu dialect from now on.

5. Handling interfaces are performed together with other
final tasks to convert to SPIR-V dialect. hal.inteface.*
and iree.placeholder ops are converted to SPIR-V global
variables, together with other loop/standard op conversions.
At the last we wrap all generated SPIR-V ops into one
SPIR-V module.

Along the way, a few cleanups and fixes are also performed.

PiperOrigin-RevId: 306033023