Don't run tf-shape-inference on CFG code.
The pass doesn't seem to handle CFG code well (didn't look in detail, but it creates verifier-invalid IR). And it seems to do enough interprocedural stuff to make our life easy-enough that we don't need a re-run after inlining.

We may want to reconsider this and add it back + extend it to support CFG in the future.

PiperOrigin-RevId: 308946191